+++ rotkehlchen/tests/unit/decoders/test_ens.py
@@ -42,7 +42,7 @@ ADDY = '0x2B888954421b424C5D3D9Ce9bB67c9bD47537d12'
 
 @pytest.mark.vcr()
 @pytest.mark.parametrize('ethereum_accounts', [[ADDY]])
-def test_mint_ens_name(database, ethereum_inquirer):
+def test_mint_ens_name(database, ethereum_inquirer, add_subgraph_api_key):  # pylint: disable=unused-argument
     tx_hash = deserialize_evm_tx_hash('0x74e72600c6cd5a1f0170a3ca38ecbf7d59edeb8ceb48adab2ed9b85d12cc2b99')  # noqa: E501
     events, decoder = get_decoded_events_of_transaction(
         evm_inquirer=ethereum_inquirer,
@@ -124,7 +124,7 @@ def test_mint_ens_name(database, ethereum_inquirer):
 
 @pytest.mark.vcr(filter_query_parameters=['apikey'])
 @pytest.mark.parametrize('ethereum_accounts', [['0x4bBa290826C253BD854121346c370a9886d1bC26']])
-def test_text_changed_old_name(database, ethereum_inquirer, ethereum_accounts):
+def test_text_changed_old_name(database, ethereum_inquirer, ethereum_accounts, add_subgraph_api_key):  # pylint: disable=unused-argument  # noqa: E501
     """
     Test that text changed for an address that is no longer
     pointed to by an ENS name does not break
@@ -175,7 +175,7 @@ def test_text_changed_old_name(database, ethereum_inquirer, ethereum_accounts):
 
 @pytest.mark.vcr(filter_query_parameters=['apikey'])
 @pytest.mark.parametrize('ethereum_accounts', [['0x4bBa290826C253BD854121346c370a9886d1bC26']])
-def test_set_resolver(database, ethereum_inquirer, ethereum_accounts):
+def test_set_resolver(database, ethereum_inquirer, ethereum_accounts, add_subgraph_api_key):  # pylint: disable=unused-argument
     tx_hex = deserialize_evm_tx_hash('0xae2cd848ce02c425bc50a8f46f8430eec32234475efb6fcff28315d2791329f6')  # noqa: E501
     evmhash = deserialize_evm_tx_hash(tx_hex)
     user_address = ethereum_accounts[0]
@@ -218,7 +218,7 @@ def test_set_resolver(database, ethereum_inquirer, ethereum_accounts):
 
 @pytest.mark.vcr()
 @pytest.mark.parametrize('ethereum_accounts', [['0xbc2E9Df6281a8e853121dc52dBc8BCc8bBE3ed0e']])
-def test_set_attribute_v2(database, ethereum_inquirer, ethereum_accounts):
+def test_set_attribute_v2(database, ethereum_inquirer, ethereum_accounts, add_subgraph_api_key):  # pylint: disable=unused-argument
     """Test that setting ens text attribute using public resolver deployed in March 2023 works"""
     tx_hex = deserialize_evm_tx_hash('0x6b354e4da21cfb06a8eb4cb5b7efd20558ae3be6a7a7c563f318e041fb3bfdd9')  # noqa: E501
     evmhash = deserialize_evm_tx_hash(tx_hex)
@@ -262,7 +262,7 @@ def test_set_attribute_v2(database, ethereum_inquirer, ethereum_accounts):
 
 @pytest.mark.vcr()
 @pytest.mark.parametrize('ethereum_accounts', [['0xA3B9E4b2C18eFB1C767542e8eb9419B840881467']])
-def test_register_v2(database, ethereum_inquirer, ethereum_accounts):
+def test_register_v2(database, ethereum_inquirer, ethereum_accounts, add_subgraph_api_key):  # pylint: disable=unused-argument
     """Test that registering an ens name using eth registar deployed in March 2023 works"""
     tx_hex = deserialize_evm_tx_hash('0x5150f6e1c76b74fa914e06df9e56577cdeec0faea11f9949ff529daeb16b1c76')  # noqa: E501
     evmhash = deserialize_evm_tx_hash(tx_hex)
@@ -322,7 +322,7 @@ def test_register_v2(database, ethereum_inquirer, ethereum_accounts):
 
 @pytest.mark.vcr()
 @pytest.mark.parametrize('ethereum_accounts', [['0xA01f6D0985389a8E106D3158A9441aC21EAC8D8c']])
-def test_renewal_with_refund_old_controller(database, ethereum_inquirer, ethereum_accounts):
+def test_renewal_with_refund_old_controller(database, ethereum_inquirer, ethereum_accounts, add_subgraph_api_key):  # pylint: disable=unused-argument  # noqa: E501
     """
     Check that if there was a refund during a renewal, the refund is subtracted from the
     spent amount. Check a refund using the old ENS registrar controller. That contract
@@ -422,7 +422,7 @@ def test_renewal_with_refund_new_controller(database, ethereum_inquirer, ethereu
 
 @pytest.mark.vcr()
 @pytest.mark.parametrize('ethereum_accounts', [['0x7277F7849966426d345D8F6B9AFD1d3d89183083']])
-def test_content_hash_changed(database, ethereum_inquirer, ethereum_accounts):
+def test_content_hash_changed(database, ethereum_inquirer, ethereum_accounts, add_subgraph_api_key):  # pylint: disable=unused-argument  # noqa: E501
     """Test that transactions changing the content hash of an ENS are properly decoded"""
     tx_hex = deserialize_evm_tx_hash('0x21fa4ef7a4c20f2548cc010ba00974632cca9e55edea4d50b3fb2c00c7f2080b')  # noqa: E501
     evmhash = deserialize_evm_tx_hash(tx_hex)
@@ -471,7 +471,7 @@ def test_content_hash_changed(database, ethereum_inquirer, ethereum_accounts):
     ('Send', ['0x4bBa290826C253BD854121346c370a9886d1bC26']),
     ('Receive', ['0x34207C538E39F2600FE672bB84A90efF190ae4C7']),
 ])
-def test_transfer_ens_name(database, ethereum_inquirer, action, ethereum_accounts):
+def test_transfer_ens_name(database, ethereum_inquirer, action, ethereum_accounts, add_subgraph_api_key):  # pylint: disable=unused-argument  # noqa: E501
     """Test that transfering an ENS name is decoded properly for all 3 cases.
 
     Owning both addresses in the transfer, only sender or only receiver
@@ -546,7 +546,7 @@ def test_transfer_ens_name(database, ethereum_inquirer, action, ethereum_account
 
 @pytest.mark.vcr()
 @pytest.mark.parametrize('ethereum_accounts', [['0x5f0eb172CaA67d45865AAd955FA77654Da33196F']])
-def test_for_truncated_labelhash(database, ethereum_inquirer, ethereum_accounts):
+def test_for_truncated_labelhash(database, ethereum_inquirer, ethereum_accounts, add_subgraph_api_key):  # pylint: disable=unused-argument  # noqa: E501
     """Test for https://github.com/rotki/rotki/issues/6597 where some labelhashes
     had their leading 0s truncated and lead to graph failures
     """
@@ -680,7 +680,7 @@ def test_vote_cast(database, ethereum_inquirer, ethereum_accounts):
 
 @pytest.mark.vcr()
 @pytest.mark.parametrize('ethereum_accounts', [[ADDY]])
-def test_set_attribute_for_non_primary_name(database, ethereum_inquirer, ethereum_accounts):
+def test_set_attribute_for_non_primary_name(database, ethereum_inquirer, ethereum_accounts, add_subgraph_api_key):  # pylint: disable=unused-argument  # noqa: E501
     """Test that setting ens text attribute for a name that is controlle by but not
     set as the primary name of the address works correctly"""
     tx_hex = deserialize_evm_tx_hash('0x07aa7d1ac61fc03f6416a25c0d6cf96f286e2ce84e9b350dd2a9a1bd6426aef2')  # noqa: E501
@@ -726,7 +726,7 @@ def test_set_attribute_for_non_primary_name(database, ethereum_inquirer, ethereu
 
 @pytest.mark.vcr()
 @pytest.mark.parametrize('ethereum_accounts', [['0xF5d90Ac6747CB3352F05BF61f48b991ACeaE28eB']])
-def test_claim_airdrop(database, ethereum_inquirer, ethereum_accounts):
+def test_claim_airdrop(database, ethereum_inquirer, ethereum_accounts, add_subgraph_api_key):  # pylint: disable=unused-argument
     tx_hex = deserialize_evm_tx_hash('0xb892797f63943dbf75e9e8a86515e9a4a964dcb6930dad10e93b526a2a648e6d')  # noqa: E501
     evmhash = deserialize_evm_tx_hash(tx_hex)
     user_address = ethereum_accounts[0]